+++ packages/@expo/cli/src/run/ios/appleDevice/protocol/AbstractProtocol.ts
@@ -6,11 +6,20 @@
  * LICENSE file in the root directory of this source tree.
  */
 import plist from '@expo/plist';
+import Debug from 'debug';
 import { Socket } from 'net';
 
+import { CommandError } from '../../../../utils/errors';
 import { parsePlistBuffer } from '../../../../utils/plist';
 
 const BPLIST_MAGIC = Buffer.from('bplist00');
+const debug = Debug('expo:apple-device:protocol');
+
+export class ProtocolClientError extends CommandError {
+  constructor(msg: string, public error: Error, public protocolMessage?: any) {
+    super(msg);
+  }
+}
 
 export type ProtocolReaderCallback = (resp: any, err?: Error) => void;
 
@@ -108,6 +117,15 @@ export abstract class ProtocolClient<MessageType = any> {
     msg: MessageType,
     callback?: (response: ResponseType, resolve: any, reject: any) => void
   ): Promise<CallbackType | ResponseType> {
+    const onError = (error: Error) => {
+      debug('Unexpected protocol socket error encountered: %s', error);
+      throw new ProtocolClientError(
+        `Unexpected protocol error encountered: ${error.message}`,
+        error,
+        msg
+      );
+    };
+
     return new Promise<ResponseType | CallbackType>((resolve, reject) => {
       const reader = this.readerFactory.create(async (response: ResponseType, error?: Error) => {
         if (error) {
@@ -119,16 +137,19 @@ export abstract class ProtocolClient<MessageType = any> {
             response,
             (value: any) => {
               this.socket.removeListener('data', reader.onData);
+              this.socket.removeListener('error', onError);
               resolve(value);
             },
             reject
           );
         } else {
           this.socket.removeListener('data', reader.onData);
+          this.socket.removeListener('error', onError);
           resolve(response);
         }
       });
       this.socket.on('data', reader.onData);
+      this.socket.on('error', onError);
       this.writer.write(this.socket, msg);
     });
   }